+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "dey",
+  "name": "bum",
   "version": "0.1.0",
   "private": true,
   "scripts": {
